Configured VagrantFile for FreeBSD
Virtualbox does not support shared folder in FreeBSD
Needed to implement NFS and make it work with vagrant
and chef.

The vagrant-berkshelf plugin does not play nice with NFS,
since the path names for berkshelf directories are too long.
Instead, for freebsd, vagrant-berkshelf plugin is disabled
and "berks install --path cookbooks" is called to move
all the berks cookbooks into the default place.

.gitignore has been added to make sure that the cookbooks
directory is not added to source control when berkshelf put them there.